Extract `collection/search.rs`
